refactor: add abstract ExpressLikeRouter to separate domain knowledge
